видаляю зайве, трішки переписую чергу для дабла(замість інта), виводжу час пошуку вершин.

+++ PriorityQueue.h
@@ -100,10 +100,10 @@ inline void PriorityQueue<T>::insert() {
 	}
 }
 
-inline void PriorityQueue<std::pair<int, int>>::insert() {
+inline void PriorityQueue<std::pair<int, double>>::insert() {
 	size_t index = sz - 1;
 	while (index > 0 and data[index - 1].second < data[index].second) {
-		std::pair<int, int> buf = data[index];
+		std::pair<int, double> buf = data[index];
 		data[index] = data[index - 1];
 		data[index - 1] = buf;
 		index--;
